Created basic jwt struct
- Holds a two cJSON structs for the header and payload. Next step is
  figuring out how to validate the token and then providing as an argument
  the struct and having it return 0 if verified and 1 if not.
- Moved out some of the utilities used in jwt.c to their own
  utils.c file. They were cluttering up the logic of jwt and weren't
  specifically jwt operations.
